requirements.txt
django==5.0.1
djangorestframework==3.14.0
python-decouple==3.8
mysqlclient==2.2.1